requirements.txt
fastapi
uvicorn
openai
pandas
scikit-learn
python-dotenv
streamlit